Adicionando PageRequest

+++ EXERCICIO-MOD 07/src/main/java/com/systemhospital/repository/HistoricRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface HistoricRepository extends JpaRepository<Historic,Integer> {
+public interface HistoricRepository extends JpaRepository<Historic,Long> {
 }
